+++ src/features/ShortFeatures.cpp
@@ -13,7 +13,7 @@ CShortFeatures::~CShortFeatures()
 }
 
 /// get feature vector for sample num
-short int* CShortFeatures::get_feature_vector(int num, int &len, bool &free)
+short int* CShortFeatures::get_feature_vector(long num, long &len, bool &free)
 {
   len=num_features ; 
   if ((num<num_vectors) && feature_matrix)
@@ -42,7 +42,7 @@ void CShortFeatures::free_feature_vector(short int* feat, bool free)
     delete[] feat ;
 } 
 
-short int* CShortFeatures::get_feature_matrix(int &num_feat, int &num_vec)
+short int* CShortFeatures::get_feature_matrix(long &num_feat, long &num_vec)
 {
   num_feat=num_features;
   num_vec=num_vectors;
