+++ src/features/ShortFeatures.h
@@ -19,12 +19,12 @@ class CShortFeatures: public CFeatures
       @param num index of feature vector
       @param len length is returned by reference
   */
-  short int* get_feature_vector(int num, int& len, bool& free);
+  short int* get_feature_vector(long num, long& len, bool& free);
   void free_feature_vector(short int* feat_vec, bool free);
   
   /// get the pointer to the feature matrix
   /// num_feat,num_vectors are returned by reference
-  short int* get_feature_matrix(int &num_feat, int &num_vec);  
+  short int* get_feature_matrix(long &num_feat, long &num_vec);  
   
   /** set feature matrix
       necessary to set feature_matrix, num_features, num_vectors, where
@@ -35,13 +35,13 @@ class CShortFeatures: public CFeatures
 protected:
   /// compute feature vector for sample num
   /// len is returned by reference
-  virtual void compute_feature_vector(int num, short int* feat)=0;
+  virtual void compute_feature_vector(long num, short int* feat)=0;
 
   /// number of features in cache
-  int num_features;
+  long num_features;
   
   /// number of vectors in cache
-  int num_vectors;
+  long num_vectors;
 
   /** chunk of memory for all the feature_vectors	
       it is aligned like 0...num_features-1 for vec0
